Register EVM client in evm/viem packages
