+++ packages/cctp-sdk/viem/src/registry.ts
@@ -1,0 +1,7 @@
+import { registerPlatformClient } from "@stable-io/cctp-sdk-definitions";
+import { ViemEvmClient } from "./viemEvmClient.js";
+
+registerPlatformClient(
+  "Evm",
+  ViemEvmClient.fromNetworkAndDomain.bind(ViemEvmClient),
+);
